Code Update
Fixed StandAtmo1976 to have density not increase with altitude, (missing
negative sign), and fixed the density functions (they kept using the
starting density instead of one in gradient before).
Removed helium mass from weight calculations because it is accounted for
in the lift calculation already.
Changed 288.15 to 288.16 in standard atmosphere table calculations to
match the actual standard